// src/app/models/documentacion/documentacion-arquitectura.model.ts
export class DocumentacionArquitectura {
    id?: any;
    areaMat?: any;
    etapaMat?: any;
    metodologiaMat?: any;
    idDoc?: any;
    publicacionMat?: string;
    firmaMat?: any;
    firmanteMat?: string;
    versionesMat?: string;
    obligatorioMat?: any;
    observacionMat?: string;
    notaMat?: string;


}